--- app/src/main/java/by/chagarin/androidlesson/CategoresFragment.java
@@ -1,240 +1,0 @@
-package by.chagarin.androidlesson;
-
-import android.app.Dialog;
-import android.app.Fragment;
-import android.app.LoaderManager;
-import android.content.AsyncTaskLoader;
-import android.content.Loader;
-import android.os.Bundle;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.app.ActionBarActivity;
-import android.support.v7.view.ActionMode;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.helper.ItemTouchHelper;
-import android.text.Editable;
-import android.text.TextUtils;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.ArrayAdapter;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Spinner;
-import android.widget.TextView;
-
-import com.melnykov.fab.FloatingActionButton;
-
-import org.androidannotations.annotations.AfterViews;
-import org.androidannotations.annotations.Click;
-import org.androidannotations.annotations.EFragment;
-import org.androidannotations.annotations.ViewById;
-
-import java.util.List;
-
-import by.chagarin.androidlesson.adapters.CategoriesAdapter;
-
-@EFragment(R.layout.fragment_categores)
-public class CategoresFragment extends Fragment {
-
-    private CategoriesAdapter categoriesAdapter;
-    private ActionModeCallback actionModeCallback = new CategoresFragment.ActionModeCallback();
-    private ActionMode actionMode;
-
-    @ViewById(R.id.categories_list_view)
-    RecyclerView recyclerView;
-
-    @ViewById(R.id.fab)
-    FloatingActionButton fab;
-
-    @ViewById
-    SwipeRefreshLayout swipeLayout;
-
-    private Spinner spinner;
-
-    @AfterViews
-    void afterView() {
-        recyclerView.setHasFixedSize(true);
-        LinearLayoutManager linearLayoutManager = new LinearLayoutManager(getActivity());
-        linearLayoutManager.setOrientation(LinearLayoutManager.VERTICAL);
-        recyclerView.setLayoutManager(linearLayoutManager);
-        fab.attachToRecyclerView(recyclerView);
-        swipeLayout.setColorSchemeColors(R.color.green, R.color.orange, R.color.blue);
-        swipeLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
-            @Override
-            public void onRefresh() {
-                loadData();
-            }
-        });
-
-        ItemTouchHelper.SimpleCallback simpleCallback = new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT) {
-            @Override
-            public boolean onMove(RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder, RecyclerView.ViewHolder target) {
-                return false;
-            }
-
-            @Override
-            public void onSwiped(RecyclerView.ViewHolder viewHolder, int direction) {
-                categoriesAdapter.removeItem(viewHolder.getAdapterPosition());
-            }
-        };
-        ItemTouchHelper itemTouchHelper = new ItemTouchHelper(simpleCallback);
-        itemTouchHelper.attachToRecyclerView(recyclerView);
-    }
-
-    //по нажатию на ФАБ запускается диалог добавления новой категории
-    @Click
-    void fabClicked() {
-        alertDialog();
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        loadData();
-    }
-
-    /**
-     * метод с помощью асинхронного загрузчика в доп потоке загружает данные из БД
-     */
-    private void loadData() {
-        getLoaderManager().restartLoader(0, null, new LoaderManager.LoaderCallbacks<List<Category>>() {
-
-            /**
-             * прозодит в бекграуде
-             */
-            @Override
-            public Loader<List<Category>> onCreateLoader(int id, Bundle args) {
-                final AsyncTaskLoader<List<Category>> loader = new AsyncTaskLoader<List<Category>>(getActivity()) {
-                    @Override
-                    public List<Category> loadInBackground() {
-                        return Category.getDataList();
-                    }
-                };
-                //важно
-                loader.forceLoad();
-                return loader;
-            }
-
-            /**
-             * в основном потоке после загрузки
-             */
-            @Override
-            public void onLoadFinished(Loader<List<Category>> loader, List<Category> data) {
-                swipeLayout.setRefreshing(false);
-                categoriesAdapter = new CategoriesAdapter(data, getActivity(), new CategoriesAdapter.CardViewHolder.ClickListener() {
-                    @Override
-                    public void onItemClick(int position) {
-                        if (actionMode != null) {
-                            toggleSelection(position);
-                        }
-                    }
-
-                    @Override
-                    public boolean onItemLongClick(int position) {
-                        if (actionMode == null) {
-                            ActionBarActivity activity = (ActionBarActivity) getActivity();
-                            actionMode = activity.startSupportActionMode(actionModeCallback);
-                        }
-                        toggleSelection(position);
-                        return true;
-                    }
-                });
-                recyclerView.setAdapter(categoriesAdapter);
-            }
-
-            @Override
-            public void onLoaderReset(Loader<List<Category>> loader) {
-
-            }
-        });
-    }
-
-    private void toggleSelection(int position) {
-        categoriesAdapter.togglePosition(position);
-        int count = categoriesAdapter.getSelectedItemsCount();
-        if (count == 0) {
-            actionMode.finish();
-        } else {
-            actionMode.setTitle(String.valueOf(count));
-            actionMode.invalidate();
-        }
-    }
-
-    /**
-     * инициируем всплывающий диалог
-     */
-    private void alertDialog() {
-        final Dialog dialog = new Dialog(getActivity());
-        dialog.setContentView(R.layout.dialog_window);
-        TextView textView = (TextView) dialog.findViewById(R.id.title);
-        spinner = (Spinner) dialog.findViewById(R.id.spinner);
-        ArrayAdapter<String> adapter = new ArrayAdapter<String>(this.getActivity(), R.layout.spinner_item, KindOfCategories.getKinds());
-        spinner.setAdapter(adapter);
-        final EditText editText = (EditText) dialog.findViewById(R.id.edit_text);
-        Button okButton = (Button) dialog.findViewById(R.id.ok_button);
-        Button cancelButton = (Button) dialog.findViewById(R.id.cancel_button);
-
-        textView.setText(R.string.categores);
-        okButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Editable text = editText.getText();
-                String kind = (String) spinner.getSelectedItem();
-                //проверка текста через Утилитный класс!!!
-                if (!TextUtils.isEmpty(text)) {
-                    new Category(text.toString(), kind).save();
-                    //прячем диалог
-                    dialog.dismiss();
-                    loadData();
-                }
-            }
-        });
-
-        cancelButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                dialog.dismiss();
-            }
-        });
-        //собственно настравиваем вид и пакезываем диалог
-        dialog.getWindow().setBackgroundDrawableResource(android.R.color.transparent);
-        dialog.show();
-    }
-
-    /**
-     * исспользуеться для создания актив мода
-     * тут описывается поведение акшн мода
-     */
-    private class ActionModeCallback implements ActionMode.Callback {
-
-        @Override
-        public boolean onCreateActionMode(ActionMode mode, Menu menu) {
-            mode.getMenuInflater().inflate(R.menu.contextual_bar, menu);
-            return true;
-        }
-
-        @Override
-        public boolean onPrepareActionMode(ActionMode mode, Menu menu) {
-            return false;
-        }
-
-        @Override
-        public boolean onActionItemClicked(ActionMode mode, MenuItem item) {
-            switch (item.getItemId()) {
-                case R.id.menu_remove:
-                    categoriesAdapter.removeItems(categoriesAdapter.getSelectedItem());
-                    mode.finish();
-                    return true;
-                default:
-                    return false;
-            }
-        }
-
-        @Override
-        public void onDestroyActionMode(ActionMode mode) {
-            categoriesAdapter.clearSelection();
-            actionMode = null;
-        }
-    }
-}
