+++ app/src/main/java/by/chagarin/androidlesson/fragments/CategoresFragment.java
@@ -1,12 +1,9 @@
-package by.chagarin.androidlesson;
+package by.chagarin.androidlesson.fragments;
 
-
+import android.app.Dialog;
 import android.app.Fragment;
 import android.app.LoaderManager;
-import android.app.SearchManager;
 import android.content.AsyncTaskLoader;
-import android.content.Context;
-import android.content.Intent;
 import android.content.Loader;
 import android.os.Bundle;
 import android.support.v4.widget.SwipeRefreshLayout;
@@ -15,94 +12,63 @@ import android.support.v7.view.ActionMode;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.helper.ItemTouchHelper;
+import android.text.Editable;
+import android.text.TextUtils;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.widget.SearchView;
+import android.view.View;
+import android.widget.ArrayAdapter;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.Spinner;
+import android.widget.TextView;
 
 import com.melnykov.fab.FloatingActionButton;
 
 import org.androidannotations.annotations.AfterViews;
-import org.androidannotations.annotations.Background;
 import org.androidannotations.annotations.Click;
 import org.androidannotations.annotations.EFragment;
-import org.androidannotations.annotations.OptionsMenu;
-import org.androidannotations.annotations.OptionsMenuItem;
 import org.androidannotations.annotations.ViewById;
-import org.androidannotations.api.BackgroundExecutor;
 
-import java.util.Date;
 import java.util.List;
 
-import by.chagarin.androidlesson.adapters.ProceedAdapter;
+import by.chagarin.androidlesson.KindOfCategories;
+import by.chagarin.androidlesson.R;
+import by.chagarin.androidlesson.adapters.CategoriesAdapter;
+import by.chagarin.androidlesson.objects.Category;
 
-@EFragment(R.layout.fragment_proceeds)
-@OptionsMenu(R.menu.menu_transactions)
-public class ProceedFragment extends Fragment {
+@EFragment(R.layout.fragment_categores)
+public class CategoresFragment extends Fragment {
 
-    private static final String TIMER_NAME = "query_timer";
-    private ProceedAdapter proceedAdapter;
-    private ActionModeCallback actionModeCallback = new ActionModeCallback();
+    private CategoriesAdapter categoriesAdapter;
+    private ActionModeCallback actionModeCallback = new CategoresFragment.ActionModeCallback();
     private ActionMode actionMode;
 
-    @OptionsMenuItem
-    MenuItem menuSearch;
-
-    @ViewById(R.id.proceeds_list)
+    @ViewById(R.id.categories_list_view)
     RecyclerView recyclerView;
 
     @ViewById(R.id.fab)
     FloatingActionButton fab;
 
-    //резинка от трусов
     @ViewById
     SwipeRefreshLayout swipeLayout;
 
-    private Proceed lastProcced;
-
-    @Override
-    public void onPrepareOptionsMenu(Menu menu) {
-        super.onPrepareOptionsMenu(menu);
-        final SearchView searchView = (SearchView) menuSearch.getActionView();
-        //делаем хинт из хмл
-        final SearchManager searchManager = (SearchManager) getActivity().getSystemService(Context.SEARCH_SERVICE);
-        searchView.setSearchableInfo(searchManager.getSearchableInfo(getActivity().getComponentName()));
-
-        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-            @Override
-            public boolean onQueryTextSubmit(String query) {
-                return false;
-            }
-
-            @Override
-            public boolean onQueryTextChange(String newText) {
-                BackgroundExecutor.cancelAll(TIMER_NAME, true);
-                filterDelayed(newText);
-                return false;
-            }
-        });
-    }
-
-    @Background(delay = 300, id = TIMER_NAME)
-    void filterDelayed(String newText) {
-        loadData(newText);
-    }
+    private Spinner spinner;
 
     @AfterViews
-    void ready() {
+    void afterView() {
         recyclerView.setHasFixedSize(true);
         LinearLayoutManager linearLayoutManager = new LinearLayoutManager(getActivity());
         linearLayoutManager.setOrientation(LinearLayoutManager.VERTICAL);
         recyclerView.setLayoutManager(linearLayoutManager);
         fab.attachToRecyclerView(recyclerView);
-        //настраиваем цветавую схесу свайпа и устанавливаем слушателя
         swipeLayout.setColorSchemeColors(R.color.green, R.color.orange, R.color.blue);
         swipeLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
             @Override
             public void onRefresh() {
-                loadData("");
+                loadData();
             }
         });
-
 
         ItemTouchHelper.SimpleCallback simpleCallback = new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT) {
             @Override
@@ -112,56 +78,40 @@ public class ProceedFragment extends Fragment {
 
             @Override
             public void onSwiped(RecyclerView.ViewHolder viewHolder, int direction) {
-                proceedAdapter.removeItem(viewHolder.getAdapterPosition());
+                categoriesAdapter.removeItem(viewHolder.getAdapterPosition());
             }
         };
         ItemTouchHelper itemTouchHelper = new ItemTouchHelper(simpleCallback);
         itemTouchHelper.attachToRecyclerView(recyclerView);
     }
 
-
+    //по нажатию на ФАБ запускается диалог добавления новой категории
     @Click
     void fabClicked() {
-        Intent intent = new Intent(getActivity(), AddProccedActivity_.class);
-        intent.putExtra(Proceed.class.getCanonicalName(), lastProcced);
-        startActivity(intent);
-        getActivity().overridePendingTransition(R.anim.from_midle, R.anim.in_midle);
+        alertDialog();
     }
 
     @Override
     public void onResume() {
         super.onResume();
-        loadData("");
-    }
-
-    private void toggleSelection(int position) {
-        proceedAdapter.togglePosition(position);
-        int count = proceedAdapter.getSelectedItemsCount();
-        if (count == 0) {
-            actionMode.finish();
-        } else {
-            actionMode.setTitle(String.valueOf(count));
-            actionMode.invalidate();
-        }
+        loadData();
     }
 
     /**
      * метод с помощью асинхронного загрузчика в доп потоке загружает данные из БД
-     *
-     * @param filter
      */
-    private void loadData(final String filter) {
-        getLoaderManager().restartLoader(0, null, new LoaderManager.LoaderCallbacks<List<Proceed>>() {
+    private void loadData() {
+        getLoaderManager().restartLoader(0, null, new LoaderManager.LoaderCallbacks<List<Category>>() {
 
             /**
              * прозодит в бекграуде
              */
             @Override
-            public Loader<List<Proceed>> onCreateLoader(int id, Bundle args) {
-                final AsyncTaskLoader<List<Proceed>> loader = new AsyncTaskLoader<List<Proceed>>(getActivity()) {
+            public Loader<List<Category>> onCreateLoader(int id, Bundle args) {
+                final AsyncTaskLoader<List<Category>> loader = new AsyncTaskLoader<List<Category>>(getActivity()) {
                     @Override
-                    public List<Proceed> loadInBackground() {
-                        return Proceed.getDataList(filter);
+                    public List<Category> loadInBackground() {
+                        return Category.getDataList();
                     }
                 };
                 //важно
@@ -173,10 +123,9 @@ public class ProceedFragment extends Fragment {
              * в основном потоке после загрузки
              */
             @Override
-            public void onLoadFinished(Loader<List<Proceed>> loader, List<Proceed> data) {
-                //отключаем свайп
+            public void onLoadFinished(Loader<List<Category>> loader, List<Category> data) {
                 swipeLayout.setRefreshing(false);
-                proceedAdapter = new ProceedAdapter(data, getActivity(), new ProceedAdapter.CardViewHolder.ClickListener() {
+                categoriesAdapter = new CategoriesAdapter(data, getActivity(), new CategoriesAdapter.CardViewHolder.ClickListener() {
                     @Override
                     public void onItemClick(int position) {
                         if (actionMode != null) {
@@ -194,31 +143,67 @@ public class ProceedFragment extends Fragment {
                         return true;
                     }
                 });
-                recyclerView.setAdapter(proceedAdapter);
-                if (data.size() != 0) {
-                    lastProcced = data.get(0);
-                } else {
-                    lastProcced = new Proceed(
-                            getString(R.string.hint_title_example),
-                            getString(R.string.hint_price_example),
-                            new Date(), "",
-                            new Category(getString(R.string.hint_category_exemple), KindOfCategories.getPlace()),
-                            new Category(getString(R.string.hint_category_exemple), KindOfCategories.getProceed()));
-                }
+                recyclerView.setAdapter(categoriesAdapter);
             }
 
             @Override
-            public void onLoaderReset(Loader<List<Proceed>> loader) {
+            public void onLoaderReset(Loader<List<Category>> loader) {
 
             }
         });
     }
 
-    /**
-     * берёт данные из БД с сортировкой от большего к меньшему
-     * @return
-     */
+    private void toggleSelection(int position) {
+        categoriesAdapter.togglePosition(position);
+        int count = categoriesAdapter.getSelectedItemsCount();
+        if (count == 0) {
+            actionMode.finish();
+        } else {
+            actionMode.setTitle(String.valueOf(count));
+            actionMode.invalidate();
+        }
+    }
 
+    /**
+     * инициируем всплывающий диалог
+     */
+    private void alertDialog() {
+        final Dialog dialog = new Dialog(getActivity());
+        dialog.setContentView(R.layout.dialog_window);
+        TextView textView = (TextView) dialog.findViewById(R.id.title);
+        spinner = (Spinner) dialog.findViewById(R.id.spinner);
+        ArrayAdapter<String> adapter = new ArrayAdapter<String>(this.getActivity(), R.layout.spinner_item, KindOfCategories.getKinds());
+        spinner.setAdapter(adapter);
+        final EditText editText = (EditText) dialog.findViewById(R.id.edit_text);
+        Button okButton = (Button) dialog.findViewById(R.id.ok_button);
+        Button cancelButton = (Button) dialog.findViewById(R.id.cancel_button);
+
+        textView.setText(R.string.categores);
+        okButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Editable text = editText.getText();
+                String kind = (String) spinner.getSelectedItem();
+                //проверка текста через Утилитный класс!!!
+                if (!TextUtils.isEmpty(text)) {
+                    new Category(text.toString(), kind).save();
+                    //прячем диалог
+                    dialog.dismiss();
+                    loadData();
+                }
+            }
+        });
+
+        cancelButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                dialog.dismiss();
+            }
+        });
+        //собственно настравиваем вид и пакезываем диалог
+        dialog.getWindow().setBackgroundDrawableResource(android.R.color.transparent);
+        dialog.show();
+    }
 
     /**
      * исспользуеться для создания актив мода
@@ -241,7 +226,7 @@ public class ProceedFragment extends Fragment {
         public boolean onActionItemClicked(ActionMode mode, MenuItem item) {
             switch (item.getItemId()) {
                 case R.id.menu_remove:
-                    proceedAdapter.removeItems(proceedAdapter.getSelectedItem());
+                    categoriesAdapter.removeItems(categoriesAdapter.getSelectedItem());
                     mode.finish();
                     return true;
                 default:
@@ -251,9 +236,8 @@ public class ProceedFragment extends Fragment {
 
         @Override
         public void onDestroyActionMode(ActionMode mode) {
-            proceedAdapter.clearSelection();
+            categoriesAdapter.clearSelection();
             actionMode = null;
         }
     }
 }
-
